lab6 complet

+++ ekelik.muhammet/I6/main.cpp
@@ -10,7 +10,7 @@ int main()
   const double mStep = 0.1;
   double m1 = 0.0;
   double m2 = 0.0;
-  size_t maxNumber = 0;
+  unsigned maxNumber = 0;
 
   std::cin >> m1 >> m2 >> maxNumber;
   if (!std::cin || m1 < -1.0 || m2 > 1.0 || m1 > m2)
